requirements.txt
# AI Trade Bot - Requirements
# Professional Trading Signal System v2.0

# =============================================================================
# CORE (pinned for stability)
# =============================================================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0

# =============================================================================
# DATA COLLECTION
# =============================================================================
ccxt>=4.0.0,<5.0.0              # Exchange API (Binance, Coinbase, etc.)
websocket-client>=1.6.0,<2.0.0  # Real-time data
requests>=2.31.0,<3.0.0         # HTTP requests

# =============================================================================
# TECHNICAL ANALYSIS
# =============================================================================
ta>=0.10.2                      # Technical indicators library
scikit-learn>=1.3.0,<2.0.0      # Preprocessing, scaling

# =============================================================================
# MACHINE LEARNING
# =============================================================================
torch>=2.0.0,<3.0.0             # PyTorch for LSTM model

# =============================================================================
# NOTIFICATIONS
# =============================================================================
plyer>=2.1.0                    # Cross-platform desktop notifications
python-telegram-bot>=20.0,<22.0 # Telegram alerts (optional)

# =============================================================================
# DASHBOARD
# =============================================================================
streamlit>=1.28.0,<2.0.0        # Web dashboard
plotly>=5.17.0,<6.0.0           # Interactive charts

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0,<2.0.0     # Environment variables
colorama>=0.4.6,<1.0.0          # Colored terminal output

# =============================================================================
# MONITORING (optional)
# =============================================================================
prometheus-client>=0.17.0       # Prometheus metrics export

# =============================================================================
# TESTING (install with: pip install -r requirements.txt[dev])
# =============================================================================
pytest>=7.0.0                   # Testing framework
